requirements.txt
flask
pandas
numpy
torch
transformers
faiss-cpu